+++ examples/resources/iosxe_interface_loopback/resource.tf
@@ -5,5 +5,4 @@ resource "iosxe_interface_loopback" "example" {
   vrf_forwarding    = "VRF1"
   ipv4_address      = "200.1.1.1"
   ipv4_address_mask = "255.255.255.255"
-  pim_sparse_mode   = true
 }