+++ examples/resources/iosxe_interface_vlan/resource.tf
@@ -6,5 +6,4 @@ resource "iosxe_interface_vlan" "example" {
   vrf_forwarding    = "VRF1"
   ipv4_address      = "10.1.1.1"
   ipv4_address_mask = "255.255.255.0"
-  pim_sparse_mode   = true
 }