+++ internal/provider/data_source_iosxe_interface_loopback_test.go
@@ -21,7 +21,6 @@ func TestAccDataSourceIosxeInterfaceLoopback(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxe_interface_loopback.test", "vrf_forwarding", "VRF1"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_loopback.test", "ipv4_address", "200.1.1.1"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_loopback.test", "ipv4_address_mask", "255.255.255.255"),
-					resource.TestCheckResourceAttr("data.iosxe_interface_loopback.test", "pim_sparse_mode", "true"),
 				),
 			},
 		},
@@ -57,7 +56,6 @@ resource "iosxe_interface_loopback" "test" {
   vrf_forwarding = "VRF1"
   ipv4_address = "200.1.1.1"
   ipv4_address_mask = "255.255.255.255"
-  pim_sparse_mode = true
   depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]
 }